+++ src/main/JaguarManager.java
@@ -8,16 +8,16 @@ import util.Time;
 
 class JaguarManager extends Entity {
 
-    private static final float INITIAL_SPAWN_INTERVAL = 3.0f;
-    private static final float SPAWN_INTERVAL_MULTIPLIER = 0.9f;
-    private static final float SPAWN_INTERVAL_DECREASE_INTERVAL = 20.0f;
+    private static final double INITIAL_SPAWN_INTERVAL = 3.0;
+    private static final double SPAWN_INTERVAL_MULTIPLIER = 0.9;
+    private static final double SPAWN_INTERVAL_DECREASE_INTERVAL = 20.0;
 
     private final Entity mHunter;
     private final HealthManager mHealthManager;
 
-    private float mCurrentSpawnInterval = INITIAL_SPAWN_INTERVAL;
-    private float mSpawnIntervalTimer = INITIAL_SPAWN_INTERVAL;
-    private float mIntervalDecreaseTimer = SPAWN_INTERVAL_DECREASE_INTERVAL;
+    private double mCurrentSpawnInterval = INITIAL_SPAWN_INTERVAL;
+    private double mSpawnIntervalTimer = INITIAL_SPAWN_INTERVAL;
+    private double mIntervalDecreaseTimer = SPAWN_INTERVAL_DECREASE_INTERVAL;
 
     JaguarManager(Entity pHunter, HealthManager pHealthManager) {
         super(buildJaguarManager());
@@ -26,8 +26,8 @@ class JaguarManager extends Entity {
     }
 
     private static IEntityData buildJaguarManager() {
-        return Entity.newDataBuilder(0.0f, 0.0f, 0.0f, 0.0f)
-                    .withColor(0.0f, 0.0f, 0.0f)
+        return Entity.newDataBuilder(0.0, 0.0, 0.0, 0.0)
+                    .withColor(0.0, 0.0, 0.0)
                     .build();
     }
 
@@ -36,37 +36,37 @@ class JaguarManager extends Entity {
         mSpawnIntervalTimer -= Time.getInstance().getDelta();
         mIntervalDecreaseTimer -= Time.getInstance().getDelta();
 
-        if (mIntervalDecreaseTimer <= 0.0f) {
+        if (mIntervalDecreaseTimer <= 0.0) {
             mCurrentSpawnInterval *= SPAWN_INTERVAL_MULTIPLIER;
             mIntervalDecreaseTimer = SPAWN_INTERVAL_DECREASE_INTERVAL;
         }
-        if (mSpawnIntervalTimer <= 0.0f) {
+        if (mSpawnIntervalTimer <= 0.0) {
             spawnJaguar();
             mSpawnIntervalTimer = mCurrentSpawnInterval;
         }
     }
 
     private void spawnJaguar() {
-        float spawnX;
-        float spawnY;
+        double spawnX;
+        double spawnY;
 
         int side = (int)(Math.random() * 4.0);
         if (side == 0) {
             // Left
-            spawnX = -(1.0f + Jaguar.WIDTH / 2.0f);
+            spawnX = -(1.0 + Jaguar.WIDTH / 2.0);
             spawnY = randomPosition(Jaguar.HEIGHT);
         } else if (side == 1) {
             // Top
             spawnX = randomPosition(Jaguar.WIDTH);
-            spawnY = 1.0f + Jaguar.HEIGHT / 2.0f;
+            spawnY = 1.0 + Jaguar.HEIGHT / 2.0;
         } else if (side == 2) {
             // Right
-            spawnX = 1.0f + Jaguar.WIDTH / 2.0f;
+            spawnX = 1.0 + Jaguar.WIDTH / 2.0;
             spawnY = randomPosition(Jaguar.HEIGHT);
         } else if (side == 3) {
             // Bottom
             spawnX = randomPosition(Jaguar.WIDTH);
-            spawnY = -(1.0f + Jaguar.HEIGHT / 2.0f);
+            spawnY = -(1.0 + Jaguar.HEIGHT / 2.0);
         } else {
             throw new InvalidStateException("Invalid jaguar spawn side " + side);
         }
@@ -74,9 +74,9 @@ class JaguarManager extends Entity {
         EntityManager.getInstance().addEntity(new Jaguar(spawnX, spawnY, mHunter, mHealthManager));
     }
 
-    private float randomPosition(float pDimension) {
-        float positionRange = 2.0f + pDimension;
-        return (float)(Math.random() * positionRange - positionRange / 2.0);
+    private double randomPosition(double pDimension) {
+        double positionRange = 2.0 + pDimension;
+        return Math.random() * positionRange - positionRange / 2.0;
     }
 
 }
